A number of small UI tweaks.
Added a few more pieces of info to the item page like size, rarity, league, etc.

A fix to make Figwheel's code reloading also redraw the site.

An interesting fix to make requesting data on an unloaded stash tab load it from
the DB.

Added the "fragments" namespace, which holds helper functions/view fragments for
rendering links to items and stashes that include a swatch of color for either
the tab's color or the item's rarity.

Added category to the stash's item list.