Fixed an issue with item links.
The problem was that loading a stash would attempt to render an item link for
each item. This loaded the item from the db, which only checks its own cache,
not the loaded stash cache, and the items weren't there, so it would trigger a
bunch of item load API calls, which overloaded localtunnel. Now, the item
subscription check the stash cache before triggering an API call.